Move `engine/fact` package to `testkit/fact`.
