+++ engine/internal/process/scope.go
@@ -8,8 +8,8 @@ import (
 	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/dogma"
 	"github.com/dogmatiq/testkit/engine/envelope"
-	"github.com/dogmatiq/testkit/engine/fact"
 	"github.com/dogmatiq/testkit/engine/panicx"
+	"github.com/dogmatiq/testkit/fact"
 	"github.com/dogmatiq/testkit/internal/location"
 )
