+++ engine/internal/process/scope_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/dogma/fixtures"
 	"github.com/dogmatiq/testkit/engine/envelope"
-	"github.com/dogmatiq/testkit/engine/fact"
 	. "github.com/dogmatiq/testkit/engine/internal/process"
+	"github.com/dogmatiq/testkit/fact"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	. "github.com/onsi/gomega/gstruct"
