+++ engine/internal/projection/controller.go
@@ -7,8 +7,8 @@ import (
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/testkit/engine/envelope"
-	"github.com/dogmatiq/testkit/engine/fact"
 	"github.com/dogmatiq/testkit/engine/panicx"
+	"github.com/dogmatiq/testkit/fact"
 )
 
 // CompactInterval is how frequently projections should be compacted.
