+++ fact/internal/logging/icon.go
