+++ fact/internal/logging/write.go
